Fix expiry, add Cloudflare auth functions, fix _turso to use static tokens
- Fix expiry: parseInt → parseFloat so sub-hour values (0.5h, 0.25h) work
- Add Cloudflare Pages Functions for login, register, user/files dashboard
- Rewrite _turso.js to use TURSO_TOKEN/TURSO_AUTH_TOKEN directly (no API token needed)
- Update Cloudflare upload to support user auth + daily limit check
- Update Cloudflare delete to support auth token (Authorization header) or deleteToken
- Add file_data/delete_token back to db.js migrations (needed for Cloudflare file storage)

+++ server/routes/files.js
@@ -58,7 +58,7 @@ router.post('/upload', upload.single('file'), async (req, res) => {
   }
 
   const short_id = nanoid(8);
-  const expiresHours = parseInt(req.body.expires_hours) || null;
+  const expiresHours = parseFloat(req.body.expires_hours) || null;
   const expires_at = expiresHours
     ? new Date(Date.now() + expiresHours * 3600 * 1000).toISOString()
     : null;
